\name{OncoSig-package}
\alias{OncoSig-package}
\alias{OncoSig}
\docType{package}
\title{
\packageTitle{OncoSig}
}
\description{
\packageDescription{OncoSig}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{OncoSig}
\packageIndices{OncoSig}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{OncoSig}

Maintainer: \packageMaintainer{OncoSig}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
