\name{OncoSigUnsup}
\alias{OncoSigUnsup}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Generatres Predictions given a precomputed Oncosig Classifier.
}
\description{
Given an OncoSig Random Forest classifier that is generated for an abritrary Oncogene-Centric Map, generates predictions for a new Oncogene/Tumor Suppressor Y, whose features are specified in the Network location. This function should be used for Oncogenes/Tumor Suppressors that do not have a gold standard for supervised learning.
}
\usage{
OncoSigUnsup(Network_location, forest_location)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Network_location}{
This is the file that specifies the features for the Oncogene/Tumor Suppressor. The feature names must be the same as the features used to generate the classifier, or an error will be reported. The format of the file is the same as passed to the function listToMatrix().
}
  \item{forest_location}{
The location of classifier, generated by OncoSigRF
}
}
\details{
This function takes in one network location and one forest location
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
#Predict KRAS Oncogene-centric map based on EGFR forest
KRAS_features="Input_data_files/LUAD/OncoSigUnsup/feature_list_KRAS.txt"
EGFR_forest="Input_data_files/LUAD/OncoSigUnsup/All_forests_EGFR.r"
results=OncoSigUnsup(KRAS_features,EGFR_forest)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
